Remove last manual relations

+++ models/upload/storable.rb
@@ -1,10 +1,7 @@
 module Upload::Storable
 
   def destroy
-    if thumbnail_id
-      thumbnail = group.uploads.find(thumbnail_id)
-      thumbnail.destroy
-    end
+    thumbnail.destroy if thumbnail
     dir = File.join('uploads', id.to_s)
     `srm -r #{dir}`
     super
